Updated README, adjusted comments on the files.

+++ customer_banking.py
@@ -11,7 +11,7 @@ def main():
     # Prompt the user to set the savings balance, interest rate, and months for the savings account.
     savings_balance = float(input("Enter your savings account balance: "))
     savings_interest = float(input("Enter your savings account interest: "))
-    savings_maturity = float(input("Enter your savings account maturity: "))
+    savings_maturity = int(input("Enter your savings account maturity: "))
 
     # Call the create_savings_account function and pass the variables from the user.
     updated_savings_balance, interest_earned = create_savings_account(savings_balance, savings_interest, savings_maturity)
@@ -22,7 +22,7 @@ def main():
     # Prompt the user to set the CD balance, interest rate, and months for the CD account.
     cd_balance = float(input("Enter your CD account balance: "))
     cd_interest = float(input("Enter your CD account interest: "))
-    cd_maturity = float(input("Enter your CD account maturity: "))
+    cd_maturity = int(input("Enter your CD account maturity: "))
 
     # Call the create_cd_account function and pass the variables from the user.
     updated_cd_balance, cd_interest_earned = create_cd_account(cd_balance, cd_interest, cd_maturity)
